+++ News/app/src/main/java/com/eugenebrusov/news/data/model/NewsItem.kt
@@ -3,7 +3,14 @@ package com.eugenebrusov.news.data.model
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
-import java.util.UUID
+import com.eugenebrusov.news.data.source.remote.models.NewsResult
+import com.eugenebrusov.news.data.source.remote.util.ApiErrorResponse
+import com.eugenebrusov.news.data.source.remote.util.ApiResponse
+import com.eugenebrusov.news.data.source.remote.util.ApiSuccessResponse
+import retrofit2.Response
+import java.text.ParseException
+import java.text.SimpleDateFormat
+import java.util.*
 
 /**
  * Immutable model class for a NewsItem.
@@ -21,4 +28,35 @@ data class NewsItem @JvmOverloads constructor(
         @ColumnInfo(name = "bodyText") var bodyText: String? = null,
         @ColumnInfo(name = "webTitle") var webTitle: String? = null,
         @ColumnInfo(name = "bylineImageUrl") var bylineImageUrl: String? = null
-)
+) {
+    companion object {
+
+        fun create(result: NewsResult): NewsItem? {
+            try {
+                val id = result.id ?: throw ParseException("Invalid news item id", 0)
+
+                val format = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'", Locale.US)
+                val webPublicationDate = format.parse(result.webPublicationDate).time
+
+                var webTitle: String? = null
+                var bylineImageUrl: String? = null
+                if (result.tags?.isNotEmpty() == true) {
+                    webTitle = result.tags[0].webTitle
+                    bylineImageUrl = result.tags[0].bylineImageUrl
+                }
+
+                return NewsItem(id = id,
+                        webPublicationDate = webPublicationDate,
+                        sectionName = result.sectionName?.toLowerCase(),
+                        headline = result.fields?.headline,
+                        trailText = result.fields?.trailText,
+                        thumbnail = result.fields?.thumbnail,
+                        bodyText = result.fields?.bodyText,
+                        webTitle = webTitle,
+                        bylineImageUrl = bylineImageUrl)
+            } catch (e: ParseException) {
+                return null
+            }
+        }
+    }
+}
